Moved repository interfaces from application layer to domain layer to make it more closer to entities.
