--- Application/Interfaces/Repositories/IRepository.cs
@@ -1,9 +1,0 @@
-﻿namespace Application.Interfaces.Repositories
-{
-    public interface IRepository<in T> where T : class
-    {
-        Task InsertAsync(T entity);
-        Task UpdateAsync(T entity);
-        Task DeleteAsync(T entity);
-    }
-}
