+++ Application/Services/UsersService.cs
@@ -1,33 +1,27 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Application.Exceptions.Types;
-using Application.Interfaces;
-using Application.Interfaces.Repositories;
+﻿using Application.Interfaces;
 using Domain.Constants;
+using Domain.Exceptions;
+using Domain.RepositoryInterfaces;
 
-namespace Application.Services
+namespace Application.Services;
+
+public class UsersService : IUsersService
 {
-    public class UsersService : IUsersService
+    private readonly IUsersRepository _userRepository;
+    private readonly ITokenService _tokenService;
+
+    public UsersService(IUsersRepository userRepository, ITokenService tokenService)
     {
-        private readonly IUsersRepository _userRepository;
-        private readonly ITokenService _tokenService;
+        _userRepository = userRepository;
+        _tokenService = tokenService;
+    }
 
-        public UsersService(IUsersRepository userRepository, ITokenService tokenService)
-        {
-            _userRepository = userRepository;
-            _tokenService = tokenService;
-        }
+    public async Task<string> AuthenticateUser(long userId)
+    {
+        var user = await _userRepository.DoesExist(userId);
+        if (user is null)
+            throw new NotFoundException($"{ApiResponseMessages.Notfound}: User ID: {userId}");
 
-        public async Task<string> AuthenticateUser(long userId)
-        {
-            var user = await _userRepository.DoesExist(userId);
-            if (user is null)
-                throw new NotFoundException($"{ApiResponseMessages.Notfound}: User ID: {userId}");
-
-            return _tokenService.GenerateJwtToken(userId, user.Name, user.AllowHistoryView);
-        }
+        return _tokenService.GenerateJwtToken(userId, user.Name, user.AllowHistoryView);
     }
 }
