+++ Domain/RepositoryInterfaces/IRepository.cs
@@ -1,0 +1,8 @@
+﻿namespace Domain.RepositoryInterfaces;
+
+public interface IRepository<in T> where T : class
+{
+    Task InsertAsync(T entity);
+    Task UpdateAsync(T entity);
+    Task DeleteAsync(T entity);
+}
