+++ Domain/RepositoryInterfaces/IUsersRepository.cs
@@ -1,0 +1,8 @@
+﻿using Domain.Entities;
+
+namespace Domain.RepositoryInterfaces;
+
+public interface IUsersRepository : IRepository<User>
+{
+    Task<User> DoesExist(long id);
+}
