+++ Infrastructure.Persistence/Repositories/UsersRepository.cs
@@ -1,37 +1,36 @@
-﻿using System;
-using Application.Interfaces.Repositories;
-using Domain.Entities;
+﻿using Domain.Entities;
+using Domain.RepositoryInterfaces;
 using Infrastructure.Persistence.Contexts;
 using Microsoft.EntityFrameworkCore;
 
-namespace Infrastructure.Persistence.Repositories
+namespace Infrastructure.Persistence.Repositories;
+
+public class UsersRepository : IUsersRepository
 {
-    public class UsersRepository : IUsersRepository
+    private readonly OfficeAccessDbContext _officeAccessDbContext;
+
+    public UsersRepository(OfficeAccessDbContext officeAccessDbContext)
     {
-        private readonly OfficeAccessDbContext _officeAccessDbContext;
+        _officeAccessDbContext = officeAccessDbContext;
+    }
 
-        public UsersRepository(OfficeAccessDbContext officeAccessDbContext)
-        {
-            _officeAccessDbContext = officeAccessDbContext;
-        }
+    public Task InsertAsync(User entity)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task InsertAsync(User entity)
-        {
-            throw new NotImplementedException();
-        }
+    public Task UpdateAsync(User entity)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task UpdateAsync(User entity)
-        {
-            throw new NotImplementedException();
-        }
+    public Task DeleteAsync(User entity)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task DeleteAsync(User entity)
-        {
-            throw new NotImplementedException();
-        }
-        public async Task<User> DoesExist(long id)
-        {
-            return await _officeAccessDbContext.Users.FirstOrDefaultAsync(x => x.Id == id);
-        }
+    public async Task<User> DoesExist(long id)
+    {
+        return await _officeAccessDbContext.Users.FirstOrDefaultAsync(x => x.Id == id);
     }
 }
